# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CAB Booking System API",
    "version": "1.0.0",
    "description": "Aggregated OpenAPI snapshot for the current CAB Booking System repo state."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Gateway"
    }
  ],
...
  "paths": {
...
    "/api/rides/{rideId}": {
      "get": {
        "tags": ["rides"],
        "summary": "Get ride details",
        "parameters": [{ "$ref": "#/components/parameters/RideIdParam" }],
        "responses": { "200": { "description": "Ride" } }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "RideIdParam": {
        "name": "rideId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    },
...
  }
}